move runner and export it (#539)
* move runner and export it

* add version str
