+++ plugin/main.go
@@ -4,31 +4,14 @@
 package main
 
 import (
-	"context"
 	"fmt"
-	"os"
 
-	"github.com/ava-labs/avalanchego/utils/logging"
-	"github.com/ava-labs/avalanchego/utils/ulimit"
 	"github.com/ava-labs/avalanchego/version"
-	"github.com/ava-labs/avalanchego/vms/rpcchainvm"
-
 	"github.com/ava-labs/subnet-evm/plugin/evm"
+	"github.com/ava-labs/subnet-evm/plugin/runner"
 )
 
 func main() {
-	printVersion, err := PrintVersion()
-	if err != nil {
-		fmt.Printf("couldn't get config: %s", err)
-		os.Exit(1)
-	}
-	if printVersion {
-		fmt.Printf("Subnet-EVM/%s [AvalancheGo=%s, rpcchainvm=%d]\n", evm.Version, version.Current, version.RPCChainVMProtocol)
-		os.Exit(0)
-	}
-	if err := ulimit.Set(ulimit.DefaultFDLimit, logging.NoLog{}); err != nil {
-		fmt.Printf("failed to set fd limit correctly due to: %s", err)
-		os.Exit(1)
-	}
-	rpcchainvm.Serve(context.Background(), &evm.VM{})
+	versionString := fmt.Sprintf("Subnet-EVM/%s [AvalancheGo=%s, rpcchainvm=%d]", evm.Version, version.Current, version.RPCChainVMProtocol)
+	runner.Run(versionString)
 }
