+++ plugin/runner/runner.go
@@ -1,0 +1,33 @@
+// (c) 2019-2020, Ava Labs, Inc. All rights reserved.
+// See the file LICENSE for licensing terms.
+
+package runner
+
+import (
+	"context"
+	"fmt"
+	"os"
+
+	"github.com/ava-labs/avalanchego/utils/logging"
+	"github.com/ava-labs/avalanchego/utils/ulimit"
+	"github.com/ava-labs/avalanchego/vms/rpcchainvm"
+
+	"github.com/ava-labs/subnet-evm/plugin/evm"
+)
+
+func Run(versionStr string) {
+	printVersion, err := PrintVersion()
+	if err != nil {
+		fmt.Printf("couldn't get config: %s", err)
+		os.Exit(1)
+	}
+	if printVersion && versionStr != "" {
+		fmt.Printf(versionStr)
+		os.Exit(0)
+	}
+	if err := ulimit.Set(ulimit.DefaultFDLimit, logging.NoLog{}); err != nil {
+		fmt.Printf("failed to set fd limit correctly due to: %s", err)
+		os.Exit(1)
+	}
+	rpcchainvm.Serve(context.Background(), &evm.VM{})
+}
